refactor(ui): refactor widget cards
