+++ ui/design/src/components/MiniInfoWidgetCard/index.tsx
@@ -2,7 +2,7 @@ import { Box, Text } from 'grommet';
 import * as React from 'react';
 import { Icon } from '../Icon';
 import { WidgetAreaCardBox } from '../EntryCard/basic-card-box';
-import { StyledButton } from './styled-widget-cards';
+import { StyledButton } from '../AppInfoWidgetCard/styled-widget-cards';
 
 export interface IMiniInfoCardProps {
   className?: string;
