+++ ui/design/src/components/TutorialWidgetCard/progress-bar.tsx
