+++ ui/design/src/components/TutorialWidgetCard/progress-node.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import styled, { css } from 'styled-components';
-import { Icon } from '../../../Icon';
+import { Icon } from '../Icon';
 
 export interface IProgressNode {
   state: NodeState;
