+++ ui/design/src/components/TutorialWidgetCard/styled-progress.ts
