+++ ui/design/src/components/TutorialWidgetCard/styled-tutorial.ts
@@ -1,5 +1,5 @@
 import styled from 'styled-components';
-import Button from '../../Button';
+import Button from '../Button';
 import { Box } from 'grommet';
 
 const StyledButton = styled(Button)`
